nyaymitra-ai/requirements.txt
python-dotenv
google-generativeai
sentence-transformers
chromadb
